src/requirements.txt
fastapi~=0.111.0
uvicorn
python-multipart
httpx
Pillow
pydantic~=2.8.2
pyyaml~=6.0.1
numpy
async-timeout
redis==5.0.7
pynvml
transformers
torch
torchvision
rq
